backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
firebase-admin==6.5.0
pandas==2.2.2
xgboost
scikit-learn
torch
torchvision
Pillow
opencv-python-headless
numpy
